+++ mechanics/src/dynamics.rs
@@ -1,4 +1,4 @@
 // Copyright (C) 2026 Yuri
 // Licensed under GPL-3.0
 
-pub mod derivatives;
+pub mod force;